cleaning and added renderer
